Crate iroh_willow

Implementation of willow

Re-exports

pub use engine::Engine;

Modules

engine

Engine for driving a willow store and synchronisation sessions.

form

Structs that allow constructing entries and other structs where some fields may be automatically filled.

interest

Types for defining synchronisation interests.

proto

Protocol data types used in willow.

rpc

session

The session module contains an implementation of the Willow General Purpose Sync Protocol (WGPS).

store

Store for entries, secrets, and capabilities used in the Willow engine.

util

Various utilities and data structures used in this crate.

Constants

ALPN

The ALPN protocol name for iroh-willow.