iroh_willow::engine
Struct Engine

pub struct Engine { /* private fields */ }

The Engine is the main handle onto a Willow store with networking.

It runs a dedicated thread for all storage operations, and a peer manager to coordinate network connections to other peers.

The engine does not establish any peer connections on its own. Synchronisation sessions can be started with Engine::sync_with_peer.

Implementations

impl Engine

pub fn client(&self) -> &MemClient

Get an in memory client to interact with the willow engine.

pub fn spawn<S: Storage>( endpoint: Endpoint, create_store: impl 'static + Send + FnOnce() -> S, accept_opts: AcceptOpts, ) -> Self

Start the Willow engine.

This needs an endpoint to connect to other peers, and a create_store closure which returns a Storage instance.

You also need to pass AcceptOpts to configure what to do with incoming connections. Its default implementation will accept all connections and run sync with all our interests.

To actually accept connections, an [Endpoint::accept] loop has to be run outside of the engine, passing all connections that match crate::net::ALPN to the engine with Engine::handle_connection.

The engine will spawn a dedicated storage thread, and the create_store closure will be called on this thread, so that the Storage does not have to be Send.

pub async fn handle_connection(&self, conn: Connection) -> Result<()>

Handle an incoming connection.

pub async fn sync_with_peer( &self, peer: NodeId, init: SessionInit, ) -> Result<IntentHandle>

Synchronises with a peer.

Will try to establish a connection to peer if there is none already, and then open a synchronisation session.

init contains the initialisation options for this synchronisation intent.

Returns an IntentHandle which receives events and can submit updates into the session.

This can freely be called multiple times for the same peer. The engine will merge the intents and make sure that only a single session is opened per peer.

pub async fn shutdown(&self) -> Result<()>

Shutdown the engine.

This will try to close all connections gracefully for up to 10 seconds, and abort them otherwise.

impl Engine

pub async fn handle_spaces_request<C: ChannelTypes<RpcService>>( self, msg: Request, chan: RpcChannel<RpcService, C>, ) -> Result<(), RpcServerError<C>>