Module iroh_willow::form

Structs that allow constructing entries and other structs where some fields may be automatically filled.

Structs

EntryForm

Creates an entry while setting some fields automatically.

Enums

AuthForm

Select which capability to use for authenticating a new entry.

EntryOrForm

Either a Entry or a EntryForm.

PayloadForm

Sources where payload data can come from.

SubspaceForm

Set the subspace either to a provided SubspaceId, or use the user authenticating the entry as subspace.

TimestampForm

Set the timestamp either to the provided Timestamp or to the current system time.