iroh_willow
Module proto

Protocol data types used in willow.

These are mostly type aliases onto [willow-rs] types, with some additional helpers.

This module also contains the crypthographic primitives for fingerprints and private area intersection.

Modules

data_model

Types for the basic data model of Willow.

grouping

Utilities for Willow’s entry groupings.

keys

Public-key crypto types for willow

meadowcap

The capability system of Willow.

pai

Primitives for Private Area Intersection

wgps

Types and helpers for the Willow General Purpose Sync protocol.