Module iroh_willow::proto::data_model

Types for the basic data model of Willow.

Modules

serde_encoding

Structs

PayloadDigest

A payload digest used in entries.

UnauthorisedWriteError

An error indicating an AuthorisationToken does not authorise the writing of this entry.

Enums

InvalidPathError

An error arising from trying to construct a invalid [Path] from valid components.

InvalidPathError2

An error arising from trying to construct a invalid Path from valid components.

Constants

DIGEST_LENGTH

The byte length of a PayloadDigest.

MAX_COMPONENT_COUNT

A natural number for limiting the number of path components.

MAX_COMPONENT_LENGTH

A natural number for limiting the length of path components.

MAX_PATH_LENGTH

A natural number max_path_length for limiting the overall size of paths.

Traits

EntryExt

Extension methods for Entry.

PathExt

Extension methods for Path.

Type Aliases

AuthorisationToken

AuthorisedEntry

An entry in a willow store.

Component

See [willow_data_model::Component].

Entry

An entry in a willow store.

NamespaceId

A type for identifying namespaces.

Path

An immutable Willow path.

SerdeWriteCapability

The capability type needed to authorize writes (serde serializable).

SubspaceId

A type for identifying subspaces.

Timestamp

A Timestamp is a 64-bit unsigned integer, that is, a natural number between zero (inclusive) and 2^64 - 1 (exclusive). Timestamps are to be interpreted as a time in microseconds since the Unix epoch.

WriteCapability

The capability type needed to authorize writes.