Module iroh_willow::proto::grouping

Structs

Area

A grouping of Entries.

AreaInArea

AreaOfInterest

A grouping of Entries that are among the newest in some store.

NamespacedRange

A three-dimensional range on a specific namespace.

NotIncluded

Point

A single point in the 3D range space.

Range

Ranges are simple, one-dimensional ways of grouping Entries.

ThreeDRange

A three-dimensional range that includes every Entry included in all three of its ranges.

Enums

RangeEnd

The end of a range, either open or closed.

SubspaceArea

Functions

path_range_end

subspace_range_end