iroh_willow::proto
Module meadowcap

The capability system of Willow.

Contains an instantiation of [meadowcap] for use in iroh-willow.

Modules

serde_encoding

Structs

ReadAuthorisation

Represents an authorisation to read an area of data in a Namespace.

Enums

AccessMode

A mode granting read or write access to some [Area].

SecretKey

Traits

IsCommunal

Maps namespace public keys to booleans, determining whether that namespace of a particular [willow_data_model::NamespaceId] is communal or owned.

Functions

is_authorised_write

is_wider_than

Returns true if self covers a larger area than other, or if covers the same area and has less delegations.

Type Aliases

FailedDelegationError

McAuthorisationToken

McCapability

McSubspaceCapability

NamespaceId

NamespacePublicKey

NamespaceSignature

ReadCapability

SubspaceCapability

UserId

UserPublicKey

UserSignature

WriteCapability