iroh_willow::proto::meadowcap
Type Alias FailedDelegationError

pub type FailedDelegationError = FailedDelegationError<MAX_COMPONENT_LENGTH, MAX_COMPONENT_COUNT, MAX_PATH_LENGTH, UserId>;

Aliased Type

enum FailedDelegationError {
    AreaNotIncluded {
        excluded_area: Area<4096, 1024, 4096, UserId>,
        claimed_receiver: UserId,
    },
    WrongSecretForUser(UserId),
}

Variants

AreaNotIncluded

The granted area of the capability we tried to delegate from did not include the given area.

Fields

excluded_area: Area<4096, 1024, 4096, UserId>

claimed_receiver: UserId

WrongSecretForUser(UserId)

The given secret did not correspond to the receiver of the existing capability we tried to delegate from.