Module iroh_willow::proto::wgps

Re-exports

pub use challenge::*;

pub use channels::*;

pub use fingerprint::*;

pub use handles::*;

pub use messages::*;

Modules

challenge

channels

fingerprint

handles

messages

Constants

MAX_PAYLOAD_SIZE

The maximum payload size limits when the other peer may include Payloads directly when transmitting Entries: when an Entry’s payload_length is strictly greater than the maximum payload size, its Payload may only be transmitted when explicitly requested.

MAX_PAYLOAD_SIZE_POWER