Module iroh_willow::proto::wgps::messages

Structs

CommitmentReveal

Complete the commitment scheme to determine the challenge for read authentication.

ControlAbsolve

Allow the other peer to reduce its total buffer capacity by amount.

ControlAnnounceDropping

The server notifies the client that it has started dropping messages and will continue to do so until it receives an Apologise message. The server must send any outstanding guarantees of the logical channel before sending a AnnounceDropping message.

ControlApologise

The client notifies the server that it can stop dropping messages on this logical channel.

ControlFreeHandle

Ask the other peer to free a resource handle.

ControlIssueGuarantee

Make a binding promise of available buffer capacity to the other peer

ControlPlead

Ask the other peer to send an ControlAbsolve message such that the receiver remaining guarantees will be target.

DataSendEntry

Transmit an AuthorisedEntry to the other peer, and optionally prepare transmission of its Payload.

DataSendPayload

Transmit some transformed Payload bytes.

DataSetMetadata

Express preferences for Payload transfer in the intersection of two AreaOfInterests.

LengthyEntry

An Entry together with information about how much of its Payload a peer holds.

PaiBindFragment

Bind data to an IntersectionHandle for performing private area intersection.

PaiReplyFragment

Finalise private set intersection for a single item.

PaiReplySubspaceCapability

Send a previously requested SubspaceCapability.

PaiRequestSubspaceCapability

Ask the receiver to send a SubspaceCapability.

ReconciliationAnnounceEntries

Prepare transmission of the LengthyEntries a peer has in a 3dRange as part of 3d range-based set reconciliation.

ReconciliationSendEntry

Transmit a LengthyEntry as part of 3d range-based set reconciliation.

ReconciliationSendFingerprint

Send a Fingerprint as part of 3d range-based set reconciliation.

ReconciliationSendPayload

Transmit some transformed Payload bytes.

ReconciliationTerminatePayload

Indicate that no more bytes will be transmitted for the currently transmitted Payload as part of set reconciliation.

SetupBindAreaOfInterest

Bind an AreaOfInterest to an AreaOfInterestHandle.

SetupBindReadCapability

Bind a ReadCapability to a CapabilityHandle.

SetupBindStaticToken

Bind a StaticToken to a StaticTokenHandle.

Enums

DataMessage

IntersectionMessage

Message

ReconciliationMessage

Type Aliases

DynamicToken

PsiGroupBytes

ReadCapability

Whereas write access control is baked into the Willow data model, read access control resides in the replication layer. To manage read access via capabilities, all peers must cooperate in sending Entries only to peers who have presented a valid read capability for the Entry. We describe the details in a capability-system-agnostic way here. To use Meadowcap for this approach, simply choose the type of valid McCapabilities with access mode read as the read capabilities.

Receiver

StaticToken

SubspaceCapability

Whenever a peer is granted a complete read capability of non-empty path, it should also be granted a corresponding subspace capability. Each subspace capability must have a single receiver (a public key of some signature scheme), and a single granted namespace (a NamespaceId). The receiver can authenticate itself by signing a collaboratively selected nonce.

SyncSignature