iroh_willow::proto::wgps
Type Alias SubspaceCapability

pub type SubspaceCapability = SerdeMcSubspaceCapability;

Whenever a peer is granted a complete read capability of non-empty path, it should also be granted a corresponding subspace capability. Each subspace capability must have a single receiver (a public key of some signature scheme), and a single granted namespace (a NamespaceId). The receiver can authenticate itself by signing a collaboratively selected nonce.

Aliased Type

struct SubspaceCapability(pub McSubspaceCapability<NamespaceId, NamespaceSignature, UserId, UserSignature>);

Fields
0: McSubspaceCapability<NamespaceId, NamespaceSignature, UserId, UserSignature>