Module iroh_willow::proto::willow

Modules

encodings

Encoding for Willow entries

Structs

AuthorisedEntry

An AuthorisedEntry is a PossiblyAuthorisedEntry for which is_authorised_write returns true.

Entry

The metadata for storing a Payload.

Path

A Path is a sequence of at most MAX_COMPONENT_COUNT many bytestrings, each of at most MAX_COMPONENT_LENGTH bytes, and whose total number of bytes is at most MAX_PATH_LENGTH.

PossiblyAuthorisedEntry

A PossiblyAuthorisedEntry is a pair of an Entry and an AuthorisationToken.

Unauthorised

Error returned for entries that are not authorised.

Enums

InvalidPath

Error returned for invalid paths.

Constants

DIGEST_LENGTH

The byte length of a PayloadDigest.

MAX_COMPONENT_COUNT

A natural number for limiting the number of path components.

MAX_COMPONENT_LENGTH

A natural number for limiting the length of path components.

MAX_PATH_LENGTH

A natural number max_path_length for limiting the overall size of paths.

Type Aliases

AuthorisationToken

Component

The type of components of a Path.

NamespaceId

A type for identifying namespaces.

PayloadDigest

A totally ordered type for content-addressing the data that Willow stores.

SubspaceId

A type for identifying subspaces.

Timestamp

A Timestamp is a 64-bit unsigned integer, that is, a natural number between zero (inclusive) and 2^64 - 1 (exclusive). Timestamps are to be interpreted as a time in microseconds since the Unix epoch.

WriteCapability

The capability type needed to authorize writes.