Module iroh_willow::proto::willow::encodings

Encoding for Willow entries

TODO: Verify that these are correct accoring to the spec! These encodings are the message bytes for authorisation signatures, so we better not need to change them again.

Structs

RelativePath

Constants

PATH_COUNT_POWER

PATH_LENGTH_POWER

PATH_LENGTH_POWER is the least natural number such that 256 ^ PATH_LENGTH_POWER ≥ MAX_COMPONENT_LENGTH. We can represent the length of any Component in path_length_power bytes. UPathLengthPower denotes the type of numbers between zero (inclusive) and 256path_length_power (exclusive).

Type Aliases

UPathCountPower

UPathLengthPower