Module iroh_willow::rpc::client

API for managing iroh spaces

iroh spaces is an implementation of the Willow protocol. The main entry point is the Client.

You obtain a Client via Iroh::spaces().

Structs

Client

Iroh Willow client.

EntryForm

Form to insert a new entry

IntentExistsError

Error returned when trying to insert a SyncHandle into a `SyncHandleSet for a peer that is already in the set.

Space

A space to store entries in.

SpaceTicket

A ticket to import and sync a space.

SyncHandle

Handle to a synchronization intent.

SyncHandleSet

Merges synchronisation intent handles into one struct.

SyncProgress

Completion state for a SyncHandle.

Type Aliases

EventReceiver

Receives events for a reconciliation intent.

UpdateSender

Sends updates into a reconciliation intent.