iroh_willow
Module session

The session module contains an implementation of the Willow General Purpose Sync Protocol (WGPS).

It exposes a few public types used to initiate sessions, and the intents module which contains handle, event and command types for controlling sessions.

Internally, this module contains the full implementation of the protocol, which is started with the run_session function (which is not public).

Modules

intents

Intents are handles onto a Willow synchronisation session.

Structs

SessionInit

Options to initialize a session.

Enums

Role

To break symmetry, we refer to the peer that initiated the synchronisation session as Alfie, and the other peer as Betty.

SessionMode

A session can either run a single reconciliation, or keep open until closed by either peer.