iroh_willow::session
Module intents

Intents are handles onto a Willow synchronisation session.

They are created with crate::Engine::sync_with_peer.

An intent receives events from the session, and can submit new interests to be synchronized.

Once all intents for a peer are complete, the session is closed.

Modules

serde_encoding

Structs

Intent

A synchronisation intent.

IntentHandle

Handle to a Intent.

SendError

Enums

Completion

Outcome of driving an intent to completion.

EventKind

Events emitted from a session for an synchronisation intent.

IntentUpdate

Updates that may be submitted from an intent into the synchronisation session.

Type Aliases

EventReceiver

IntentId

UpdateSender