iroh_willow::store::traits
Trait EntryStorage

pub trait EntryStorage:
    EntryReader
    + Clone
    + Debug
    + 'static {
    type Reader: EntryReader;
    type Snapshot: EntryReader + Clone;

    // Required methods
    fn reader(&self) -> Self::Reader;
    fn snapshot(&self) -> Result<Self::Snapshot>;
    fn ingest_entry(
        &self,
        entry: &AuthorisedEntry,
        origin: EntryOrigin,
    ) -> Result<bool>;
    fn subscribe_area(
        &self,
        namespace: NamespaceId,
        area: Area,
        params: SubscribeParams,
    ) -> impl Stream<Item = StoreEvent> + Unpin + 'static;
    fn resume_subscription(
        &self,
        progress_id: u64,
        namespace: NamespaceId,
        area: Area,
        params: SubscribeParams,
    ) -> impl Stream<Item = StoreEvent> + Unpin + 'static;
}

Storage for entries.

Required Associated Types

type Reader: EntryReader

type Snapshot: EntryReader + Clone

Required Methods

fn reader(&self) -> Self::Reader

fn snapshot(&self) -> Result<Self::Snapshot>

fn ingest_entry( &self, entry: &AuthorisedEntry, origin: EntryOrigin, ) -> Result<bool>

Ingest a new entry.

Returns true if the entry was ingested, and false if the entry was not ingested because a newer entry exists.

fn subscribe_area( &self, namespace: NamespaceId, area: Area, params: SubscribeParams, ) -> impl Stream<Item = StoreEvent> + Unpin + 'static

Subscribe to events concerning entries included by an [AreaOfInterest], returning a producer of StoreEvents which occurred since the moment of calling this function.

If ignore_incomplete_payloads is true, the producer will not produce entries with incomplete corresponding payloads. If ignore_empty_payloads is true, the producer will not produce entries with a payload_length of 0.

fn resume_subscription( &self, progress_id: u64, namespace: NamespaceId, area: Area, params: SubscribeParams, ) -> impl Stream<Item = StoreEvent> + Unpin + 'static

Attempt to resume a subscription using a progress ID obtained from a previous subscription, or return an error if this store implementation is unable to resume the subscription.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.