iroh_willow::util::channel
Function inbound_channel

pub fn inbound_channel<T: Decoder>(
    max_buffer_size: usize,
) -> (Writer, Receiver<T>)

Create a new channel with a byte Writer on the transmit side and a message Receiver on the receive side.

This is used for data incoming from the network: The networking code copies received data into the channel, and the application code processes the messages parsed by the Decoder from the data in the channel.