Function iroh_willow::util::pipe::chunked_pipe

pub fn chunked_pipe(chunk_size: usize) -> (PipeWriter, PipeReader)

In-memory local-io async pipe between a [AsyncStreamWriter] and a [Stream] of [Bytes].

The pipe maintains a shared in-memory buffer of chunk_size

PipeWriter is a [AsyncStreamWriter] that writes into the shared buffer.

PipeReader is [Stream] that emits [Bytes] of chunk_size length. The last chunk may be smaller than chunk_size.

The pipe is closed once either the reader or the writer are dropped. If the reader is dropped, subsequent writes will fail with io::ErrorKind::BrokenPipe.