iroh_willow::util::channel
Function outbound_channel

pub fn outbound_channel<T: Encoder>(
    max_buffer_size: usize,
    guarantees: Guarantees,
) -> (Sender<T>, Reader)

Create a new channel with a message Sender on the transmit side and a byte Reader on the receive side.

This is used for data sent from the application into the network: The application code queues messages for sending, and the networking code consumes a bytes stream of the messages encoded with Encoder.

Optionally the channel can be assigned a limited number of Guarantees. If limited, a total limit of sendable bytes will be respected, and no further sends can happen once it is exhausted. The amount of guarantees can be raised with Sender::add_guarantees.